ci: GitHub Actions: `dorny/paths-filter@v3` update
In openfoodfacts/openfoodfacts-server#11624 the original plan was to
add some negative matching blobs using `paths-filter`, but that did
not function as intended.

As a result of that, we switched to use `tj-actions/paths-filter`
instead, in pull requests #11630 and #11644.

The reason `changed-files` did not work as expected is/was the lack
of dorny/paths-filter#224 - a feature only available from v3.0.2
onwards.

This changeset enables support for the `predicate-quantifier` setting
and configures it for the negative-blob matches on the `code`
fileset.